refactor: Remove misc for now
