+++ server/src/app.module.ts
@@ -9,7 +9,7 @@ import { PostgresConnectionOptions } from 'typeorm/driver/postgres/PostgresConne
 import { UsersModule } from './users/users.module'
 import { AuthModule } from './auth/auth.module'
 import { PlayerSessionModule } from './player-session/player-session.module'
-import { DecksModule } from './decks/decks'
+import { DecksModule } from './decks/decks.module'
 
 @Module({
   imports: [
